requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-discord
apache-airflow-providers-postgres
apache-airflow-providers-slack
apache-airflow-providers-pagerduty
apache-airflow-providers-snowflake
awswrangler
moto
polars